win: fixes for Windows x64
Mostly size_t <-> unsigned int warnings, but I also had a mistake in
PROCESS_BASIC_INFORMATION, the pids are 32-on-32 and 64-on-64.

The Windows build is still x86 until https://codereview.chromium.org/981333002/.
I don't think I'll bother maintaining the x86 build for now, though we will probably
need it for x86 OSs in the future. It should be straightforward to revive it once we
need it, and have bots to support it.

R=mark@chromium.org
BUG=crashpad:1

Review URL: https://codereview.chromium.org/983103004

+++ client/crash_report_database_win.cc
@@ -390,7 +390,7 @@ void Metadata::Read() {
 
   base::CheckedNumeric<uint32_t> records_size =
       base::CheckedNumeric<uint32_t>(header.num_records) *
-      sizeof(MetadataFileReportRecord);
+      static_cast<uint32_t>(sizeof(MetadataFileReportRecord));
   if (!records_size.IsValid()) {
     LOG(ERROR) << "record size out of range";
     return;
