+++ util/net/http_multipart_builder.cc
@@ -46,7 +46,8 @@ std::string GenerateBoundaryString() {
   for (int index = 0; index < 32; ++index) {
     const char kCharacters[] =
         "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-    int random_value = base::RandInt(0, strlen(kCharacters) - 1);
+    int random_value =
+        base::RandInt(0, static_cast<int>(strlen(kCharacters)) - 1);
     boundary_string += kCharacters[random_value];
   }
   boundary_string += "---";
